Add dom/NodeInlines.h for inlines functions of Node class https://bugs.webkit.org/show_bug.cgi?id=288953
Reviewed by Ryosuke Niwa.

Inline functions of Node class were defined in CharacterData.h,
ElementInlines.h, DocumentInlines.h and Element.cpp. Added
NodeInlines.h to gather.

Added BoundaryPointInlines.h, EditingInlines.h, PositionInlines.h and
RangeBoundaryPointInlines.h for inline functions which need
NodeInlines.h.

* Source/WebCore/Headers.cmake:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/accessibility/AccessibilityNodeObject.cpp:
* Source/WebCore/accessibility/AccessibilityObject.cpp:
* Source/WebCore/accessibility/atspi/AccessibilityObjectTextAtspi.cpp:
* Source/WebCore/dom/BoundaryPoint.h:
(WebCore::makeBoundaryPointAfterNodeContents): Deleted.
* Source/WebCore/dom/BoundaryPointInlines.h: Added.
(WebCore::makeBoundaryPointAfterNodeContents):
* Source/WebCore/dom/CharacterData.h:
(WebCore::Node::length const): Deleted.
* Source/WebCore/dom/DocumentInlines.h:
(WebCore::Node::scriptExecutionContext const): Deleted.
(WebCore::Node::protectedScriptExecutionContext const): Deleted.
(WebCore::Node::opaqueRoot const): Deleted.
(WebCore::Node::protectedDocument const): Deleted.
* Source/WebCore/dom/Element.cpp:
(WebCore::Node::setTabIndexState): Deleted.
* Source/WebCore/dom/ElementInlines.h:
(WebCore::Node::hasAttributes const): Deleted.
(WebCore::Node::attributesMap const): Deleted.
(WebCore::Node::parentElement const): Deleted.
(WebCore::Node::protectedParentElement const): Deleted.
* Source/WebCore/dom/FragmentDirectiveGenerator.cpp:
* Source/WebCore/dom/FragmentDirectiveRangeFinder.cpp:
* Source/WebCore/dom/Node.cpp:
(WebCore::Node::scriptExecutionContext const):
* Source/WebCore/dom/Node.h:
* Source/WebCore/dom/NodeInlines.h: Added.
(WebCore::Node::protectedScriptExecutionContext const):
(WebCore::Node::opaqueRoot const):
(WebCore::Node::protectedDocument const):
(WebCore::Node::hasAttributes const):
(WebCore::Node::attributesMap const):
(WebCore::Node::parentElement const):
(WebCore::Node::protectedParentElement const):
(WebCore::Node::setTabIndexState):
(WebCore::Node::length const):
* Source/WebCore/dom/Position.cpp:
* Source/WebCore/dom/Position.h:
(WebCore::lastPositionInNode): Deleted.
* Source/WebCore/dom/PositionInlines.h: Added.
(WebCore::lastPositionInNode):
* Source/WebCore/dom/PositionIterator.cpp:
* Source/WebCore/dom/Range.cpp:
* Source/WebCore/dom/RangeBoundaryPoint.h:
(WebCore::RangeBoundaryPoint::setToAfterContents): Deleted.
* Source/WebCore/dom/RangeBoundaryPointInlines.h: Added.
(WebCore::RangeBoundaryPoint::setToAfterContents):
* Source/WebCore/dom/SimpleRange.cpp:
* Source/WebCore/editing/ApplyBlockElementCommand.cpp:
* Source/WebCore/editing/ApplyStyleCommand.cpp:
* Source/WebCore/editing/CompositeEditCommand.cpp:
* Source/WebCore/editing/DeleteSelectionCommand.cpp:
* Source/WebCore/editing/Editing.cpp:
* Source/WebCore/editing/Editing.h:
(WebCore::lastPositionInOrAfterNode): Deleted.
* Source/WebCore/editing/EditingInlines.h: Added.
(WebCore::lastPositionInOrAfterNode):
* Source/WebCore/editing/Editor.cpp:
* Source/WebCore/editing/IndentOutdentCommand.cpp:
* Source/WebCore/editing/InsertLineBreakCommand.cpp:
* Source/WebCore/editing/InsertListCommand.cpp:
* Source/WebCore/editing/InsertTextCommand.cpp:
* Source/WebCore/editing/ReplaceSelectionCommand.cpp:
* Source/WebCore/editing/TextManipulationController.cpp:
* Source/WebCore/editing/TypingCommand.cpp:
* Source/WebCore/editing/VisibleSelection.cpp:
* Source/WebCore/editing/VisibleUnits.cpp:
* Source/WebCore/editing/markup.cpp:
* Source/WebCore/html/HTMLTextFormControlElement.cpp:
* Source/WebCore/rendering/RenderBox.cpp:
* Source/WebCore/rendering/RenderObject.cpp:
* Source/WebKit/Shared/EditingRange.cpp:
* Source/WebKit/WebProcess/InjectedBundle/DOM/InjectedBundleNodeHandle.cpp:
* Source/WebKit/WebProcess/WebPage/ios/WebPageIOS.mm:
* Source/WebKitLegacy/mac/WebView/WebFrame.mm:

Canonical link: https://commits.webkit.org/291488@main
