+++ Source/WebCore/dom/ElementInlines.h
@@ -50,29 +50,6 @@ inline unsigned Element::findAttributeIndexByName(const AtomString& name, bool s
     return elementData()->findAttributeIndexByName(name, shouldIgnoreAttributeCase);
 }
 
-inline bool Node::hasAttributes() const
-{
-    auto* element = dynamicDowncast<Element>(*this);
-    return element && element->hasAttributes();
-}
-
-inline NamedNodeMap* Node::attributesMap() const
-{
-    if (auto* element = dynamicDowncast<Element>(*this))
-        return &element->attributesMap();
-    return nullptr;
-}
-
-inline Element* Node::parentElement() const
-{
-    return dynamicDowncast<Element>(parentNode());
-}
-
-inline RefPtr<Element> Node::protectedParentElement() const
-{
-    return parentElement();
-}
-
 inline const Element* Element::rootElement() const
 {
     if (isConnected())
