+++ Source/WebCore/editing/DeleteSelectionCommand.cpp
@@ -28,7 +28,7 @@
 
 #include "Document.h"
 #include "DocumentMarkerController.h"
-#include "Editing.h"
+#include "EditingInlines.h"
 #include "Editor.h"
 #include "EditorClient.h"
 #include "ElementInlines.h"
