+++ Source/WebCore/editing/ReplaceSelectionCommand.cpp
@@ -40,8 +40,8 @@
 #include "DataTransfer.h"
 #include "Document.h"
 #include "DocumentFragment.h"
-#include "Editing.h"
 #include "EditingBehavior.h"
+#include "EditingInlines.h"
 #include "ElementIteratorInlines.h"
 #include "EventNames.h"
 #include "FilterOperations.h"
