+++ Source/WebKit/Shared/EditingRange.cpp
@@ -26,8 +26,10 @@
 #include "config.h"
 #include "EditingRange.h"
 
+#include <WebCore/BoundaryPointInlines.h>
 #include <WebCore/FrameSelection.h>
 #include <WebCore/LocalFrame.h>
+#include <WebCore/RangeBoundaryPointInlines.h>
 #include <WebCore/TextIterator.h>
 #include <WebCore/VisibleUnits.h>
